+++ reference/examples/gctb_2.0_tutorial/gwas/plink_submit_combination.sh
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f4c1128220f9eb53415ba6eaad102c99f9fc56e71649100b0a73852c7268ef03
+size 410