+++ reference/examples/gctb_2.0_tutorial/ldm/makeldm_full.sh
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:cdfb2a291991c9715b291d7a75e5d46e8d95a06c30a09cd7132e824f96f610d9
+size 644