+++ reference/examples/gctb_2.0_tutorial/ldm/makeldm_full_mcpu.sh
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d74e3c9fc0f0e9d2f3e6c6716abc068338621e173214987b92eb8b699195c90b
+size 819