+++ CHANGELOG.md
@@ -31,6 +31,7 @@ If MD5 sum is not listed for a certain release then it means that the container 
   ec089544b13d3eb39f13728f8584dcde  saige.sif   (update to SAIGE v0.44.6.5)
   627734a5c74c94bd69453d0366aced5a  r.sif       (new packages)
   ```
+- add ``gctb_2.0_tutorial.zip`` (under ``reference/examples/gctb_2.0_tutorial``)
 
 ## [1.0.0] - 2020-10-20